split navbar into 2 components navbar and header
