+++ src/styles/styles.scss
@@ -1,5 +1,6 @@
 @import "./settings";
-@import "./components/navigation";
+@import "./components/header";
+@import "./components/navbar";
 @import "./pages/home";
 @import "./pages/beers";
 @import "./pages/profile";
